rename rsalibnum.py to number_theory.py and all callings
