+++ attacks/multi_keys/common_factors.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 from attacks.abstract_attack import AbstractAttack
-from lib.rsalibnum import gcd, ProductTree
+from lib.number_theory import gcd, ProductTree
 from lib.keys_wrapper import PrivateKey
 from gmpy2 import mul
