+++ attacks/single_key/fermat_numbers_gcd.py
@@ -4,7 +4,7 @@
 from attacks.abstract_attack import AbstractAttack
 from tqdm import tqdm
 from lib.keys_wrapper import PrivateKey
-from lib.rsalibnum import gcd
+from lib.number_theory import gcd
 
 
 class Attack(AbstractAttack):
