+++ attacks/single_key/fibonacci_gcd.py
@@ -4,7 +4,7 @@
 from attacks.abstract_attack import AbstractAttack
 from tqdm import tqdm
 from lib.keys_wrapper import PrivateKey
-from lib.rsalibnum import gcd, fib
+from lib.number_theory import gcd, fib
 
 
 class Attack(AbstractAttack):
