+++ attacks/single_key/mersenne_pm1_gcd.py
@@ -4,7 +4,7 @@
 from attacks.abstract_attack import AbstractAttack
 from tqdm import tqdm
 from lib.keys_wrapper import PrivateKey
-from lib.rsalibnum import gcd, ilog2
+from lib.number_theory import gcd, ilog2
 
 
 class Attack(AbstractAttack):
