+++ attacks/single_key/system_primes_gcd.py
@@ -5,7 +5,7 @@ from attacks.abstract_attack import AbstractAttack
 from tqdm import tqdm
 from lib.keys_wrapper import PrivateKey
 from lib.system_primes import load_system_consts
-from lib.rsalibnum import gcd
+from lib.number_theory import gcd
 
 
 class Attack(AbstractAttack):
