+++ lib/keys_wrapper.py
@@ -10,7 +10,7 @@ from lib.crypto_wrapper import PKCS1_OAEP
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.backends import default_backend
 from lib.conspicuous_check import privatekey_check
-from lib.rsalibnum import powmod, invmod, invert
+from lib.number_theory import powmod, invmod, invert
 
 # from gmpy2 import invert
