+++ lib/rsa_attack.py
@@ -9,7 +9,7 @@ from lib.utils import print_results
 from lib.fdb import send2fdb
 from lib.crypto_wrapper import bytes_to_long, long_to_bytes
 import inspect
-from lib.rsalibnum import is_prime, isqrt, gcd
+from lib.number_theory import is_prime, isqrt, gcd
 import traceback
