+++ sage/small_crt_exp.sage
@@ -1,5 +1,5 @@
 from sage.libs.ntl.ntl_ZZ_pX import ntl_ZZ_pContext, ntl_ZZ_pX
-from lib.rsalibnum import powmod
+from lib.number_theory import powmod
 
 def poly_fast_ntl(ctx, f, xs):
     # Fast multipoint evaulation from Modern Computer Algebra 3rd edition 10.1
